chore: move Selector out of componeonts
