+++ src/pages/common/FilterBar/FilterBar.tsx
@@ -6,7 +6,7 @@ import { TAGS_MOCK } from 'src/mocks/tags.mock'
 
 import { Box, Flex } from 'theme-ui'
 import { Button } from 'oa-components'
-import Selector from 'src/components/Selector'
+import Selector from './Selector'
 import { logger } from 'src/logger'
 import { Link } from 'react-router-dom'
